pyproject.toml
[project]
name = "ragdemo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12.5"
dependencies = [
    "annotated-types==0.7.0",
    "anyio==4.10.0",
    "certifi==2025.8.3",
    "cffi==1.17.1",
    "charset-normalizer==3.4.2",
    "cryptography==45.0.5",
    "distro==1.9.0",
    "docx==0.2.4",
    "dotenv==0.9.9",
    "faiss-cpu==1.11.0.post1",
    "fastapi==0.116.1",
    "filelock==3.18.0",
    "fsspec==2025.7.0",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httpx==0.28.1",
    "idna==3.10",
    "jinja2==3.1.6",
    "jiter==0.10.0",
    "joblib==1.5.1",
    "lxml==6.0.0",
    "markupsafe==3.0.2",
    "mpmath==1.3.0",
    "networkx==3.5",
    "numpy==2.3.2",
    "nvidia-cublas-cu12==12.6.4.1",
    "nvidia-cuda-cupti-cu12==12.6.80",
    "nvidia-cuda-nvrtc-cu12==12.6.77",
    "nvidia-cuda-runtime-cu12==12.6.77",
    "nvidia-cudnn-cu12==9.5.1.17",
    "nvidia-cufft-cu12==11.3.0.4",
    "nvidia-cufile-cu12==1.11.1.6",
    "nvidia-curand-cu12==10.3.7.77",
    "nvidia-cusolver-cu12==11.7.1.2",
    "nvidia-cusparse-cu12==12.5.4.2",
    "nvidia-cusparselt-cu12==0.6.3",
    "nvidia-nccl-cu12==2.26.2",
    "nvidia-nvjitlink-cu12==12.6.85",
    "nvidia-nvtx-cu12==12.6.77",
    "ollama==0.5.1",
    "openai==1.98.0",
    "packaging==25.0",
    "pdfminer-six==20250506",
    "pdfplumber==0.11.7",
    "pillow==11.3.0",
    "pycparser==2.22",
    "pydantic==2.11.7",
    "pydantic-core==2.33.2",
    "pypdfium2==4.30.0",
    "python-docx==1.2.0",
    "python-dotenv==1.1.1",
    "regex==2025.7.34",
    "requests==2.32.4",
    "scikit-learn==1.7.1",
    "scipy==1.16.1",
    "setuptools==80.9.0",
    "sniffio==1.3.1",
    "starlette==0.47.2",
    "sympy==1.14.0",
    "threadpoolctl==3.6.0",
    "tiktoken==0.9.0",
    "torch==2.7.1",
    "tqdm==4.67.1",
    "triton==3.3.1",
    "typing-extensions==4.14.1",
    "typing-inspection==0.4.1",
    "urllib3==2.5.0",
]
